requirements.txt
Django>=5.0
djangorestframework>=3.15.0
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.1
ipython                               